Move test template to the correct directory - fixes #1520
