+++ eclipse-collections-code-generator/src/main/resources/test/set/immutable/immutablePrimitiveEmptySetTest.stg
@@ -22,7 +22,6 @@ import java.util.NoSuchElementException;
 import org.eclipse.collections.api.set.primitive.Immutable<name>Set;
 import org.eclipse.collections.impl.factory.primitive.<name>Bags;
 import org.eclipse.collections.impl.factory.primitive.<name>Sets;
-import org.eclipse.collections.impl.set.mutable.primitive.AbstractImmutable<name>HashSetTestCase;
 import org.eclipse.collections.impl.test.Verify;
 import org.junit.Assert;
 import org.junit.Test;
