+++ eclipse-collections-code-generator/src/main/resources/test/set/immutable/immutablePrimitiveKeySetTest.stg
@@ -5,7 +5,7 @@ isTest() ::= "true"
 
 skipBoolean() ::= "true"
 
-targetPath() ::= "org/eclipse/collections/impl/map/mutable/primitive"
+targetPath() ::= "org/eclipse/collections/impl/map/immutable/primitive"
 
 fileName(primitive) ::= "Immutable<primitive.name>MapKeySetTest"
 
@@ -16,11 +16,12 @@ class(primitive) ::= <<
 body(type, wrapperName, name) ::= <<
 <copyright()>
 
-package org.eclipse.collections.impl.map.mutable.primitive;
+package org.eclipse.collections.impl.map.immutable.primitive;
 
 import org.eclipse.collections.api.set.primitive.<name>Set;
 import org.eclipse.collections.api.set.primitive.Immutable<name>Set;
-import org.eclipse.collections.impl.set.mutable.primitive.AbstractImmutable<name>HashSetTestCase;
+import org.eclipse.collections.impl.map.mutable.primitive.<name>ByteHashMap;
+import org.eclipse.collections.impl.set.immutable.primitive.AbstractImmutable<name>HashSetTestCase;
 import org.junit.Assert;
 
 /**
